Added Project 2 folder (begining project)
Humble beginings: here we mark the start of our journey back into Jquery and that wonderful world of programming. With a week to the deadline will our young protagonist complete on time, or will he be stopped by the dasterdly distractions that lie all around him. Stay tuned to find out.