Project 2: Update 1
Episode one: The phantom bug menaces my code.
The first objective has been achieved:
I've made an image go off screen & return back to it's starting position.
The next objective is already being worked on  the image changes after the first click to a new image.
However not all is well, my issue is that for some reason the animation stacks so on the second click function toggle happens twice and thrice and four time ect... so until this is solved my code is not good. Will Yan succeed in solving the problem or will he succumb to this herculean bug.