Project 2: Update 2
Episode 2: Revenge of the Coder.
In this episode our protagonist has solved the issue of the stacking animations the bug has been destroyed but how? The solution was adding stopPropogation to the end of the Toggle function however although it solves the issue it mysteriously is registering in the console as a reference error.
The images now change off screen, and are pulled randomly from an array.
And now the third objective is being accomplished adding images to the array depending on the amount of clicks.