Project 3: Border, Increase WorldSize, Removed distance check
Border: Now at the border of the map,  there are border tiles which when stepped on, bring you back to center stage. This is done through overlapCheck(); and NewTiles is called which reloads the map.
Increase WorldSize: For better effect, larger play area, no bleed.
Removed distance check: I'd be lying if I said I didn't feel a tinge of irony and despair at removing this from my code but after several hours of working on a distance based wrapper, i realized the most effective solution was a border based one which would allow the player to explore the most.
So thats what I did, and for my efforts I got some serious lack of sleep :+1: